requirements.txt
Pillow==7.2.0
pyzmq==19.0.2
